:art: :tractor: Move prompto commands to their own files
